use auto-updated manifest
